Split classes into individual modules
All public apis are still collected and available through
lsst.validate.base; there should be no outward API change.
This change makes much easier to maintain validate_base.
